suppress or fix more clippy errors

+++ src/parquet2/src/read/column/stream.rs
@@ -20,7 +20,7 @@ where
 
     let mut chunk = vec![];
     chunk.try_reserve(length as usize)?;
-    reader.take(length as u64).read_to_end(&mut chunk).await?;
+    reader.take(length).read_to_end(&mut chunk).await?;
     Result::Ok(chunk)
 }
